// src/app/views/actions/auth/auth.action-type.tsx
enum AUTH_ACTION_TYPES {
    SET_UID = "SET_UID",

    REGISTER_REQUEST = "CLIENTS_REGISTER_REQUEST",
    REGISTER_SUCCESS = "CLIENTS_REGISTER_SUCCESS",
    REGISTER_FAILURE = "CLIENTS_REGISTER_FAILURE",

    LOGIN_REQUEST = "CLIENTS_LOGIN_REQUEST",
    LOGIN_SUCCESS = "CLIENTS_LOGIN_SUCCESS",
    LOGIN_FAILURE = "CLIENTS_LOGIN_FAILURE",

    LOGOUT = "CLIENTS_LOGOUT"
};

interface ISetUid {
    type: AUTH_ACTION_TYPES.SET_UID;
    uid: string;
}

type LOGIN_ACTION = ISetUid;

export {
    AUTH_ACTION_TYPES,
    LOGIN_ACTION
}